Fix wrong displaying attributes when no fg given
In a `prytheme` file, when you don't provide a foreground color and use
only attributes (for exapmple, `symbol : (b)`), you always get `maroon`
color (color `1`). This is not in order, because a user expects that
color wouldn't change. The expected result would be "symbol uses default
terminal color, but with bold font".

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>