Fix pry-theme crash when no `ENV['host_os']`
Fix issue #4 (Can't convert nil into String (TypeError))

-----------------------------------------------------------------------
Backtrace provided by KINGSABRI:

  % pry --simple-prompt
  /var/lib/gems/1.8/gems/pry-theme-0.0.7/lib/pry-theme.rb:30:in `join':
    can't convert nil into String (TypeError)
-----------------------------------------------------------------------

We slightly change the approach in this commit. Without trying to guess
whether we on Mac OS or GNU/Linux, we just forcibly use `$HOME/.pry`
directory in every case but Windows.

I thought that `EVN['host_os']` is a variable, which we can rely on,
but, as practice shows, sometimes it weirdly goes back on.

Reported-by: KINGSABRI <king.sabri@gmail.com>.
Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>