Refactor copying of default themes to theme dir
Move routine operations into the Helper and distribute them into
methods:

  * theme_file_version detects a version of a `.prytheme` file. The
    detection relies on the fact that a version of a theme is defined on
    the fourth line of the file. This is not very robust solution, but
    in the current state of things it works perfectly, as I don't plan
    to change the structure of the Pry Themes;

  * default_theme and local_theme methods just save a fair amount of
    typing.

Now, the most important part: I finally thought up of a real useful
application for version attribute of `.prytheme` files. In this commit I
use it for keeping installed themes up to date. Let's say I have
an installed theme, which version is 1. This is a default theme, which
means that it comes with the standard distribution of the gem. Now, in
the new gem version, I made a couple of tweaks for that theme and
changed its version to 4. With a regular update of the gem the old code
wouldn't have reacted on that situation and would have ignored the fact
that that theme has been updated. So with a new gem version, instead of
the fresh theme we would've gotten nothing, because the code assumes
that if the file has been installed once, then we should not touch it
anymore. Which was, obviously, wrong approach.

The new approach is based on versioning and takes into account `version`
attribute and *really* updates old themes.

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>