Fix unwanted fall back to the default theme
Before this commit, when a user enters a name of nonexistent theme, say,
`pry-theme yaddayadda`, Pry Theme replies with the message that
`yaddayadda` is not installed and falls back to the default theme. The
latter isn't what we really want:

  [17] (pry) main: 0> pry-theme classic
  classic.prytheme doesn't exist
  Using classic theme

Moreover, it states that it switches to "classic" theme, but in reality,
it switches to the default one.

In this commit Pry Theme falls back only in the case when a user defines
nonexistent theme in `Pry.config.theme`.

Also, as the side effect, change Helper, so it's a module_function now.

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>