Improve `--edit` command
First of all, move Theme.pathify_theme method to Helper. Then, add an
argument support for `-e` command. For example:

  pry(main)> pry-theme -e pry-classic

This will open `pry-classic` theme for editing.

Also, add a support for theme creation mechanism. For example:

  pry(main)> pry-theme -e +new_theme

The "+" mark indicates that we want to create a new theme. The command
will create a new file in theme directory and fill it with data from the
template (with some hints).

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>