Color256: refactor "#find_from_fixnum"
I have to refactor it, because Rubinius in 1.9 mode is unable to find a
colour in certain conditions (no idea, what's going on, but it does fail
every time: https://travis-ci.org/kyrylo/pry-theme/jobs/5096452).

The problem is the splat operator in block arguments. I couldn't
reproduce the issue outside of pry-theme, so screw it.